# Cache device on TensorImpl; clean up TensorImpl constructors.