--- a/c10/core/Storage.h
+++ b/c10/core/Storage.h
@@ -41,11 +41,12 @@ struct C10_API Storage {
   // that can be temporarily created with Caffe2 APIs. See the note on top of
   // TensorImpl.h for details.
   static Storage create_legacy(at::Device device, caffe2::TypeMeta data_type) {
+    auto allocator = GetAllocator(device.type());
     return Storage(c10::make_intrusive<StorageImpl>(
             data_type,
             0,
-            at::DataPtr(nullptr, device),
-            GetAllocator(device.type()),
+            allocator->allocate(0), // materialize a non-default Device.
+            allocator,
             true));
   }