--- a/torch/csrc/autograd/variable.cpp
+++ b/torch/csrc/autograd/variable.cpp
@@ -22,7 +22,7 @@
 namespace torch {
 namespace autograd {
 Variable::Impl::Impl(at::Tensor data, std::unique_ptr<Variable::AutogradMeta> autograd_meta, bool requires_grad, Edge gradient_edge)
-    : TensorImpl(data.type_id(), data.dtype(), /*allocator=*/nullptr, /* is variable */ true),
+    : TensorImpl(data.type_id(), data.dtype(), data.device(), /* is variable */ true),
       data_(std::move(data)) {
   autograd_meta->grad_fn_ = std::move(gradient_edge.function);
   autograd_meta->requires_grad_ = false;
@@ -101,10 +101,6 @@ const at::Storage& Variable::Impl::storage() const {
 
 int64_t Variable::Impl::storage_offset() const {
   return data_.storage_offset();
-}
-
-int64_t Variable::Impl::get_device_slow() const {
-  return data_.get_device();
 }
 
 std::shared_ptr<Function> Variable::grad_accumulator() const {