@gchanan 2019-03-28
Actually it looks like the benchmarks are broken for previous versions.  But getting the device is faster!

Old:
```
Benchmark                                 Time           CPU Iterations
BM_TensorImplGetDevice                    3 ns          3 ns  279652190
```

New:
```
Benchmark                                 Time           CPU Iterations
BM_TensorImplGetDevice                    2 ns          2 ns  331110583

```