@gchanan 2019-03-29
I incorporated https://github.com/pytorch/pytorch/pull/18605 into here.  I'm not sure if it's correct, but I didn't check if it's possible to use ghstack to stack this on top of a new PR, so I just threw it in here to get test signal.