@gchanan 2019-04-04
closing in favor of https://github.com/pytorch/pytorch/pull/18833.